requirement.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
passlib
python-dotenv
pandas
streamlit
requests
plotly
python-multipart
pyotp
qrcode[pil]
Pillow